[FC-7-10]: update fetchContactList and searchContact actions, redusers, actionCreators
