+++ 7-8-9-10-NodeJS+React/src/components/AddContactPage.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import ContactForm from './ContactForm';
-import { addContact } from '../actions';
+import { addContact } from '../actions/types';
 
 const propTypes = {
   contactList: PropTypes.arrayOf(PropTypes.shape({
