+++ 7-8-9-10-NodeJS+React/src/index.js
@@ -1,16 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { createStore } from 'redux';
-import contactList from './data';
+import { applyMiddleware, createStore, compose } from 'redux';
+import thunk from 'redux-thunk';
 import rootReducer from './reducers';
+import { fetchContacts } from './actions';
 import App from './components/app';
 
 const store = createStore(
   rootReducer,
-  { contactList },
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
+  compose(applyMiddleware(thunk),
+    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()),
 );
+
+store.dispatch(fetchContacts());
 
 ReactDOM.render(
   <Provider store={store}>
