+++ 7-8-9-10-NodeJS+React/src/reducers/searchValue.js
@@ -1,6 +1,8 @@
+import { SEARCH_CONTACT } from '../actions/types';
+
 const searchValue = (state = '', action) => {
   switch (action.type) {
-    case 'SEARCH_CONTACT':
+    case SEARCH_CONTACT:
       return action.searchValue;
     default:
       return state;
